chore: reorganize config dir
