+++ backend/app/config/env_settings/cors_config.py
