+++ backend/tests/test_config_manager.py
@@ -1,10 +1,10 @@
 import importlib
 
 import pytest
-from app.config.config import BackendConfiguration, settings
-from app.config.cors_config import CORSConfig
-from app.config.database_config import DatabaseConfig
-from app.config.logger_config import LoggerConfig
+from app.config.env_settings.backend_config import BackendConfiguration, settings
+from app.config.env_settings.cors_config import CORSConfig
+from app.config.env_settings.database_config import DatabaseConfig
+from app.config.env_settings.logger_config import LoggerConfig
 from pydantic import ValidationError
 from pydantic_settings import SettingsConfigDict
